+++ src/components/Hero.tsx
@@ -7,7 +7,7 @@ import Image from 'next/image'
 import ImageSrc from '@public/image.jpg'
 import Link from 'next/link'
 import { PageInfo } from '@types.d'
-import { urlFor } from '@lib/imageUrResolver'
+import { urlFor } from '@lib/imageUrlResolver'
 
 type Props = {
   pageInfo: PageInfo
@@ -36,6 +36,7 @@ export default function Hero({ pageInfo }: Props) {
         loading='eager'
         height={128}
         width={128}
+        sizes='(max-width: 600px) 100vw, (max-width: 900px) 50vw, 33.3vw'
       />
 
       <div className='space-y-2 relative'>
