+++ src/components/Skill.tsx
@@ -3,7 +3,7 @@
 import React from 'react'
 import { motion } from 'framer-motion'
 import { Skill } from '../../typings'
-import { urlFor } from '@lib/imageUrResolver'
+import { urlFor } from '@lib/imageUrlResolver'
 import Image from 'next/image'
 
 type Props = {
@@ -24,7 +24,7 @@ function Skill({ directionLeft, skill }: Props) {
       className='group relative flex cursor-pointer'
       title={skill?.title}
     >
-      <div className='rounded-full border border-gray-500 object-cover w-20 h-20 md:w-28 md:h-28  xl:w-32 xl:h-32 filter group-hover:grayscale duration-300 ease-in-out items-center justify-center flex relative'>
+      <div className='rounded-full border border-gray-500 object-cover size-20 md:size-28 xl:size-32 filter group-hover:grayscale duration-300 ease-in-out items-center justify-center flex relative'>
         <Image
           alt={skill.title}
           src={
@@ -33,10 +33,11 @@ function Skill({ directionLeft, skill }: Props) {
           }
           width={56}
           height={56}
+          sizes='(max-width: 600px) 100vw, (max-width: 900px) 50vw, 33.3vw'
         />
       </div>
 
-      <div className='bg-primary absolute rounded-full w-24 h-24 md:w-28 md:h-28 xl:w-32 xl:h-32 items-center flex justify-center opacity-0 hover:opacity-80 transition-opacity duration-200 ease-in-out'>
+      <div className='bg-primary absolute rounded-full size-24 md:size-28 xl:size-32 items-center flex justify-center opacity-0 hover:opacity-80 transition-opacity duration-200 ease-in-out'>
         <div className='bg-[#242424] rounded-full w-4/5 h-4/5 items-center flex justify-center '>
           <p className='text-base xl:text-xl font-bold text-primary opacity-full'>
             {skill.progress}
