+++ src/lib/imageUrlResolver.ts
@@ -6,4 +6,3 @@ import { config } from './config'
  * Read more: https://www.sanity.io/docs/image-url
  **/
 export const urlFor = (source: Image) =>
-  createImageUrlBuilder(config).image(source)
